move breakpoint stuff to detect-breakpoint
