+++ static/src/javascripts/projects/commercial/modules/comment-adverts.ts
@@ -1,5 +1,5 @@
 import { adSizes, createAdSlot } from '@guardian/commercial-core';
-import { getCurrentBreakpoint } from 'lib/detect-viewport';
+import { getCurrentBreakpoint } from 'lib/detect-breakpoint';
 import config from '../../../lib/config';
 import fastdom from '../../../lib/fastdom-promise';
 import { mediator } from '../../../lib/mediator';
