+++ static/src/javascripts/projects/commercial/modules/header-bidding/utils.ts
@@ -1,7 +1,10 @@
 import { createAdSize } from '@guardian/commercial-core';
 import { isString } from '@guardian/libs';
 import { once } from 'lodash-es';
-import { getCurrentTweakpoint, matchesBreakpoints } from 'lib/detect-viewport';
+import {
+	getCurrentTweakpoint,
+	matchesBreakpoints,
+} from 'lib/detect-breakpoint';
 import config from '../../../../lib/config';
 import { pbTestNameMap } from '../../../../lib/url';
 import {
