+++ static/src/javascripts/projects/commercial/modules/messenger/passback.ts
@@ -2,7 +2,7 @@ import { adSizes } from '@guardian/commercial-core';
 import type { RegisterListener } from '@guardian/commercial-core';
 import { log } from '@guardian/libs';
 import { breakpoints } from '@guardian/source-foundations';
-import { getCurrentBreakpoint } from 'lib/detect-viewport';
+import { getCurrentBreakpoint } from 'lib/detect-breakpoint';
 import fastdom from '../../../../lib/fastdom-promise';
 import { adSlotIdPrefix } from '../dfp/dfp-env-globals';
